Adding missing postal_fee_id to payment_items, edge case handling
* Postal fee ID was not added in the products initial migration
as product ID was. This migration is adding it in case it's still
missing.
* Because not all columns could be available by default in payments
form, product-related columns were conditioned and accessed only
if they're present. This is a temporary solution as the whole
form is planned to be rewritten to comply with dataproviders and
flow where external modules can extend it.

remp/crm#1118