Use public_name instead of email (ProductsModule)
Email is still present in CheckoutFormFactory. This is fine; it's used
to log in and authenticators have access to username while
authenticating user.

Refactoring use of email as user identification used through system.
In future it will be either email, phone number or unique nickname
filled in `users.public_name`.

remp/crm#1145