Fix routing issues with shop
When shop was supposed to link to other modules and it was run
on the custom domain, routing didn't work as expected as "show"
pattern matched <module>/<presenter>/<action> pattern.

I made "show" pattern to be more specific so when not matched,
default routing kicks in and links work correctly.

remp/novydenik#633