Hotfix purchase of more item pcs than in stock
Previous implementation checks if there is at least one item in stock
when rendering cart/checkout. But this isn't correct in case someone
wants to purchase multiple pieces of one item. We need to check pieces
of item in cart against stock.

This is just hotifx.

- Correctly informs customer about insufficient number of pieces in
  stock when adding to cart.
- Removes product from cart when openning cart and checking out order
  (with flash message about products out of stock).

Proper implementation should process number of ordered items against
stock and remove only number of pieces that are not in stock - not whole
product.

remp/crm#1641