Disable plus button if no more items are in stock
Multiple pieces of products can be purchased in shop. In cart there are
minus and plus buttons which allow update number of items per product.

If number of items in stock is same as number of items in cart, we do
not allow adding more items. Button is disabled and tooltip is
displayed.

Note: `ajaxComplete` is required to reload tooltip initialization,
      otherwise we lose options `html: true` after ajax redraws button.

remp/crm#1641