Inform user if product was not stored into distribution center
Added catching of `ProductSaveEventDistributionException`
into `ProductsFormFactory` to inform user that product was
not stored into distribution center.

This exception has to be thrown from inside `ProductSaveEvent` handler
when your module is creating / updating product in distribution center.

KikaModule now throws this exception if creation/update was not
successful.

remp/crm#1804