Fix occasional missing postal fee notice
The notice was caused by switching of country and immediate
form submission. When the country is changed, the list of
available postal fees changes too. But occasionally people
submit the form without waiting for the list change and
backend evaluates this as if no postal fee was selected.

Previously this caused notices and since PHP 8.1 also some
deprecations.

Commit fixes the behavior, displays postal fee in the checkout
summary only if the backend evaluated it as actually selected.
Also postalFeeChange! handler is triggered only when it was
actually changed. The previous implementation triggered it
even when nothing was selected and value of the first postal
fee was sent to the backend.

remp/crm#2706